Create TCS3200ArduinoUno.ino
This is the main file for the Arduino Uno. This takes the value from the TCS3200 and then instructs the servos where to move the Sweet. In this configuration this sensor was quite reliable but do note that the background material or any ambient light will result in erroneous reading and a lot of trouble shooting